utility functions moved utils.cc
